feat(web): switch new web to default home
Primary:
- Move new UI routes underneath "/"
- Move old UI routes underneath "/old"
- Harden application and changeset creation logic for new UI
  - Ensure that we do not create duplicates
  - Ensure that swapping between UIs does not affect state
- Remove ApplicationView "on leave" logic in order to maintain
  application and changeset state

Misc:
- Ensure AssetPalette is reactive to the ApplicationId rather than just
  the first one ("1")
- Delete unused OldAppSwitch atom
