+++ app/web/src/organisms/Application/ApplicationView.vue
@@ -30,53 +30,13 @@
       -->
     </div>
     <!-- this one is extra -->
-
-    <SiModal v-model="unsavedChangesModalShow" name="unsavedChanges">
-      <template #title> Unsaved Changes</template>
-      <template #body>
-        <div class="flex flex-col items-center w-full h-full mb-2">
-          <div class="text-base font-normal text-red-500">
-            You have unsaved changes!
-          </div>
-          <div class="text-sm text-white">Are you sure you want to leave?</div>
-        </div>
-      </template>
-      <template #buttons>
-        <SiButton
-          size="sm"
-          label="leave"
-          class="mx-1"
-          :icon="null"
-          kind="cancel"
-          @click="leave"
-        />
-        <SiButton
-          size="sm"
-          label="stay"
-          class="mx-1"
-          :icon="null"
-          kind="save"
-          @click="stay"
-        />
-      </template>
-    </SiModal>
   </div>
 </template>
 
 <script setup lang="ts">
-import { onUnmounted, ref, watch } from "vue";
-
+import { watch } from "vue";
 import Editor from "@/organisms/Editor.vue";
-import SiModal from "@/molecules/SiModal.vue";
-import SiButton from "@/atoms/SiButton.vue";
-import {
-  onBeforeRouteLeave,
-  RouteLocationNormalized,
-  useRouter,
-} from "vue-router";
-import { ChangeSetService } from "@/service/change_set";
 import { refFrom, untilUnmounted } from "vuse-rx";
-import { switchToHead } from "@/service/change_set/switch_to_head";
 import _ from "lodash";
 import { ApplicationService } from "@/service/application";
 import { setIfError } from "@/service/global_error";
@@ -109,42 +69,4 @@ watch(
 const application = refFrom<Component | null>(
   ApplicationService.currentApplication(),
 );
-
-const editMode = refFrom(ChangeSetService.currentEditMode());
-const navDestination = ref<RouteLocationNormalized | null>(null);
-const unsavedChangesModalShow = ref<boolean>(false);
-const reallyLeave = ref<boolean>(false);
-
-const router = useRouter();
-const leave = () => {
-  if (navDestination.value) {
-    unsavedChangesModalShow.value = false;
-    reallyLeave.value = true;
-    switchToHead();
-    router
-      .push(navDestination.value)
-      .catch((err) => console.log("route error", { err }));
-  }
-};
-
-const stay = () => {
-  unsavedChangesModalShow.value = false;
-  navDestination.value = null;
-};
-
-onBeforeRouteLeave((to, _from) => {
-  if (editMode.value && reallyLeave.value === false) {
-    if (!_.isNull(to)) {
-      navDestination.value = to;
-      unsavedChangesModalShow.value = true;
-      return false;
-    }
-  }
-  return true;
-});
-
-onUnmounted(() => {
-  ApplicationService.clearCurrentApplication();
-  ChangeSetService.switchToHead();
-});
 </script>
